# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cinema Booking System API",
        "description": "API documentation for Cinema Booking System",
        "contact": {
            "email": "support@cinemat.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{protocol}://{host}",
            "description": "API Server",
            "variables": {
                "protocol": {
                    "enum": [
                        "http",
                        "https"
                    ],
                    "default": "http"
                },
                "host": {
                    "default": "localhost:8089"
                }
            }
        }
    ],
    "paths": {
...
        "/api/seats/hold": {
            "post": {
                "tags": [
                    "Seats"
                ],
                "summary": "Giữ chỗ ghế",
                "operationId": "a9556fe67198ea52bba14ef8faa4b8d5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "showtime_id",
                                    "seat_id"
                                ],
                                "properties": {
                                    "showtime_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "seat_id": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Giữ chỗ thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Ghế đã được giữ chỗ"
                    },
                    "400": {
                        "description": "Lỗi validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/seats/release": {
            "post": {
                "tags": [
                    "Seats"
                ],
                "summary": "Giải phóng ghế",
                "operationId": "40f6cf57161e03a7fddb223b284be3ce",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "showtime_id",
                                    "seat_id"
                                ],
                                "properties": {
                                    "showtime_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "seat_id": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Giải phóng thành công"
                    },
                    "404": {
                        "description": "Ghế không được giữ chỗ"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/seats/showtime/{showtimeId}/held": {
            "get": {
                "tags": [
                    "Seats"
                ],
                "summary": "Lấy danh sách ghế đã được giữ chỗ",
                "operationId": "534562dbdc40e27d10f066ed4956871c",
                "parameters": [
                    {
                        "name": "showtimeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/seats/check-status": {
            "post": {
                "tags": [
                    "Seats"
                ],
                "summary": "Kiểm tra trạng thái giữ chỗ ghế",
                "operationId": "61f88a763807607e4e862a640e1a4264",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "showtime_id",
                                    "seat_id"
                                ],
                                "properties": {
                                    "showtime_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "seat_id": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "is_held": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "hold_info": {
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}